Changes to f32 floats
Updates the codebase to use f32 floats instead of f64. This change aims to improve performance and reduce memory usage on embedded targets

+++ examples/ez.rs
@@ -20,7 +20,7 @@ fn main() -> Result<(), fasteval3::Error> {
     //    |            many built-in functions: print, int, ceil, floor, abs, sign, log, round, min, max, sin, asin, ...
     //    standard binary operators
 
-    assert!((val - 1.23).abs() < f64::EPSILON);
+    assert!((val - 1.23).abs() < f32::EPSILON);
 
     Ok(())
 }
