+++ examples/ns_callback.rs
@@ -2,7 +2,7 @@
 
 fn main() -> Result<(), fasteval3::Error> {
     let mut num_lookups = 0;
-    let mut cb = |name: &str, _args: Vec<f64>| -> Option<f64> {
+    let mut cb = |name: &str, _args: Vec<f32>| -> Option<f32> {
         num_lookups += 1;
         match name {
             "x" => Some(2.0),
@@ -11,7 +11,7 @@ fn main() -> Result<(), fasteval3::Error> {
     };
 
     let val = fasteval3::ez_eval("x * (x + 1)", &mut cb)?;
-    assert!((val - 6.0).abs() < f64::EPSILON);
+    assert!((val - 6.0).abs() < f32::EPSILON);
     assert_eq!(num_lookups, 2); // Notice that 'x' was looked-up twice.
 
     Ok(())
